LPD-27477 Reorganize code according to command renames
