--- modules/_node-scripts/.gitignore
@@ -1,1 +1,0 @@
-!/build
