+++ modules/_node-scripts/bundle/amd/util/getNamespacedPackageName.mjs
