+++ modules/_node-scripts/bundle/amd/util/splitProjectExport.mjs
