+++ modules/_node-scripts/bundle/amd/writeExportBridges.mjs
