+++ modules/_node-scripts/bundle/amd/writeMainBridge.mjs
