+++ modules/_node-scripts/bundle/amd/writeManifestJson.mjs
