+++ modules/_node-scripts/bundle/amd/writePackageJson.mjs
