+++ modules/_node-scripts/bundle/esbuild/bundleJavaScriptMain.mjs
