+++ modules/_node-scripts/bundle/esbuild/getEntryPoint.mjs
