+++ modules/_node-scripts/bundle/esbuild/getExternals.mjs
