+++ modules/_node-scripts/bundle/esbuild/getPathPrefix.mjs
