+++ modules/_node-scripts/tsconfig/baseTsconfig.mjs
