+++ modules/_node-scripts/tsconfig/writeProjectTsconfig.mjs
@@ -10,8 +10,8 @@ import {
 	SRC_PATH,
 	SRC_TSCONFIG_PATH,
 	getRootDir,
-} from '../../util/constants.mjs';
-import sortObjectKeys from '../../util/sortObjectKeys.mjs';
+} from '../util/constants.mjs';
+import sortObjectKeys from '../util/sortObjectKeys.mjs';
 import baseTsconfig from './baseTsconfig.mjs';
 
 export default async function writeProjectTsconfig(
